recipe-recommender/requirements.txt
streamlit
fastapi
uvicorn
pandas
numpy
sentence-transformers
faiss-cpu
openai
python-dotenv
